Support Dialog Boxes (#457)
* Separated `Window` and `Pot`. (#456)

* Supported setting debug window status. (#456)

* Bug fixed: the callbacks attached to the variable that remain after the widget is destroyed cause errors that it cannot find the widget. (#456)

* Supported warning popups. (#456)

* Bug fixed. (#456)

* Bug fixed: unmatched types. (#456)

* Bug fixed: wrong vertical position. (#456)

+++ leads/data_persistence/analyzer/processor.py
@@ -189,7 +189,7 @@ class Processor(object):
             return (dt := t - self._lap_start_time), (
                 ds := mileage - self._lap_start_mileage), 3600000 * ds / dt if dt else 0
 
-        def shared_post(duration: float, distance: float, avg_speed: float) -> None:
+        def shared_post(duration: int, distance: float, avg_speed: float) -> None:
             if self._max_lap_duration is None or duration > self._max_lap_duration:
                 self._max_lap_duration = duration
             if self._max_lap_distance is None or distance > self._max_lap_distance:
